+++ Probability/WeightedInteger.cs
@@ -8,15 +8,15 @@ namespace Probability
     // Weighted integer distribution using alias method.
     public sealed class WeightedInteger : IDiscreteDistribution<int>, IWeightedDistribution<int>
     {
-        private readonly List<int> weights;
+        private readonly List<long> weights;
         private readonly IDistribution<int>[] distributions;
 
-        public static IDiscreteDistribution<int> Distribution(params int[] weights) =>
-            Distribution((IEnumerable<int>)weights);
+        public static IDiscreteDistribution<int> Distribution(params long[] weights) =>
+            Distribution((IEnumerable<long>)weights);
 
-        public static IDiscreteDistribution<int> Distribution(IEnumerable<int> weights)
+        public static IDiscreteDistribution<int> Distribution(IEnumerable<long> weights)
         {
-            List<int> w = weights.ToList();
+            List<long> w = weights.ToList();
             if (w.Any(x => x < 0))
                 throw new ArgumentException();
             if (!w.Any(x => x > 0))
@@ -25,23 +25,23 @@ namespace Probability
                 return Singleton<int>.Distribution(0);
             if (w.Count == 2)
                 return Bernoulli.Distribution(w[0], w[1]);
-            int gcd = weights.GCD();
+            long gcd = weights.GCD();
             for (int i = 0; i < w.Count; i += 1)
                 w[i] /= gcd;
             return new WeightedInteger(w);
         }
 
-        private WeightedInteger(IEnumerable<int> weights)
+        private WeightedInteger(IEnumerable<long> weights)
         {
             this.weights = weights.ToList();
-            int s = this.weights.Sum();
+            long s = this.weights.Sum();
             int n = this.weights.Count;
             this.distributions = new IDistribution<int>[n];
-            var lows = new Dictionary<int, int>();
-            var highs = new Dictionary<int, int>();
+            var lows = new Dictionary<int, long>();
+            var highs = new Dictionary<int, long>();
             for (int i = 0; i < n; i += 1)
             {
-                int w = this.weights[i] * n;
+                long w = this.weights[i] * n;
                 if (w == s)
                     this.distributions[i] = Singleton<int>.Distribution(i);
                 else if (w < s)
@@ -55,11 +55,11 @@ namespace Probability
                 lows.Remove(low.Key);
                 var high = highs.First();
                 highs.Remove(high.Key);
-                int lowNeeds = s - low.Value;
+                long lowNeeds = s - low.Value;
                 this.distributions[low.Key] =
                     Bernoulli.Distribution(low.Value, lowNeeds)
                       .Select(x => x == 0 ? low.Key : high.Key);
-                int newHigh = high.Value - lowNeeds;
+                long newHigh = high.Value - lowNeeds;
                 if (newHigh == s)
                     this.distributions[high.Key] =
                       Singleton<int>.Distribution(high.Key);
@@ -73,7 +73,7 @@ namespace Probability
         public IEnumerable<int> Support() =>
           Enumerable.Range(0, weights.Count).Where(x => weights[x] != 0);
 
-        public int Weight(int i) =>
+        public long Weight(int i) =>
             0 <= i && i < weights.Count ? weights[i] : 0;
 
         double IWeightedDistribution<int>.Weight(int t) => this.Weight(t);
